Add PreHandle to API/HTTPEasy/HTTP
PreHandle is an optional method that is called immediately upon receiving the HTTP request, before authentication
and before rate limit checks. This method allows servers to provide early handling of a request before any
processing happens.

The value of the error is not used, only if an error or nil was returned. If an error is returned then no more
processing is performed. It is assumed that a response will have been written to w.

If nil is returned then the request will continue normally, no status should have been written to w. Any headers
added may be overwritten by the handle.